pyproject.toml
[project]
name = "elsa"
version = "0.1.0"
description = "Technical demo for ELSA (Sr. SWE position)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "arrow>=1.3.0",
  "fastapi[standard]>=0.115.6",
  "loguru>=0.7.3",
  "nats-py>=2.9.0",
  "pre-commit>=4.0.1",
  "redis>=5.2.1",
  "rich>=13.9.4",
  "ruff>=0.8.3",
  "websocket>=0.2.1",
]
